Adds desktop wallet lock time setting
The lock time (default 3 minutes) can now be adjusted under the advanced settings menu.
The minimum lock time is set to 1 minute.

Under the Idle component, the handleEvent() function is now debounced and will only be called if no event happened after 500 milliseconds. This function was previously called a lot of times (for example for each single
mouseevent).

Added an instruction under the desktop README on how to get the app working under Windows as the keytar module must be rebuilt with electron-rebuild.

Added a new translation key called 'lockScreenTimeout' for the new setting (I've added en and de).

Fixed various typos in comments of the touched components.

Signed-off-by: Luca Moser <moser.luca@gmail.com>