Adds auto. node switching functionality
The src/shared/libs/iota/index.js singleton object now has a proxy
object injected into it, which automatically intercepts calls to api
function invocations (via ES6 Proxy). Before executing a iota library
function, the interceptor checks whether the current specified node
returns information for getNodeInfo calls, if yes, the origin function
is called, if not, the node is swapped by a new random node of the
available ones and then the origin function is executed. Only functions
which use a callback as their last argument are intercepted, as
synchronous iota library functions do not.

The approach with the Proxy was used, because the api function
invocations are all over the place, so that it becomes non trivial to
really check before hand, whether the node will work.

The user automatically gets an alert when the node was auto. switched.
Because a separate Worker is used in the desktop version for certain
calls to the API, messaging between the main app and the Worker are set
in place, to ensure that both iota library instances (they are
isolated, app/Worker) have the same state up on mutation.

The interceptor makes the assumption, that the call will succeed when
the node was auto. switched. If that wouldn't be the case, the user is
very likely offline and offline state should be handled in another way.

Added an AutoNodeSwitching HOC which handles the auto. switching
adjustments. Additionally, the Node HOC also now can mutate the settings
which controls the auto. switching. The Index component of the desktop
app is now wrapped in the AutoNodeSwitching HOC.

Added additional translation sentences for the auto. node switching
alert.

Mobile wallet developers can use the shared HOCs to implement the same
thing.

Signed-off-by: Luca Moser <moser.luca@gmail.com>