Fixes custom node add issue (attachToTangle)
Constructs a fake response object in case the node doesn't supply a JSON
reponse we normally expect when attachToTangle isn't available.
Additionally this commit removes a non existent function call and adds
the parameter 'addingCustomNode' on setFullNode when a node is added, so
that the node is actually added to the custom node list.