Expose Argon2JNI as a native module
Argon2JNI (https://github.com/marcoh00/argon2jni) is not well maintained. The only possible usage option is to import it as a module. To avoid shipping the source for this lib, a fork was made to add jitpack support (https://github.com/laumair/argon2jni). The fork updates the build configuration to make it work with jitpack. The library integration is further extended to a NativeModule implementation for RN mobile.