Mobile: Fix unrecognised password issue
An handled exception generates an alert "Unrecognised password" on login screen (mobile). Following are the steps to reproduce the issue:

1- Complete onboarding
2- Close the application while the account is loading for the first time
3- Reopen the application
4- Type in password

After 4) it will generate an alert "Unrecognised password" when in fact, the password is correct.

This commit fixes the issue by handling the exception that lead to this incorrect error alert.