fix: do not perform quorum when it's explicitly turned off
We do not perform quorum in all areas of the application and therefore have used param variables for determining when and when not to perform quorum. With the node settings update (where we have allowed users to disable quorum completely), there existed an issue where the quorum was still happening despite deactivation.

This commit fixes the issue and ensures that on deactivation, quorum will not be performed in any area of the application.

Related issue: https://github.com/iotaledger/trinity-wallet/issues/1739