fix: store complete token details in store (#2482)
We store token (object) details in redux store if a user does not opt for storing his/her payment card details on MoonPay servers. However, on desktop we were only storing the token id (and not the complete token object) hence it was giving undefined values on Review Purchase screen.

This commit fixes the issue by storing the complete token details in store.